Got seperate file for api's working
